kmeix: cast to uint64_t when using time_t value

+++ src/modules/kemix/kemix_mod.c
@@ -23,6 +23,7 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 #include <time.h>
 
 #include "../../core/sr_module.h"
@@ -952,7 +953,7 @@ static sr_kemi_xval_t* ki_kx_get_nhuri(sip_msg_t *msg)
  */
 static int ki_kx_get_timestamp(sip_msg_t *msg)
 {
-	return (int)time(NULL);
+	return (int)(uint64_t)time(NULL);
 }
 
 /**
